ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tensorflow>=2.20.0
scikit-learn==1.3.2
xgboost==2.0.2
Pillow==10.1.0
numpy>=1.26.0
pandas==2.1.3
python-multipart==0.0.6
pydantic==2.5.0
slowapi==0.1.9
python-dotenv==1.0.0
